requirements.txt
langchain-community
openai
supabase
tiktoken
python-dotenv